+++ src/ui/pprof.rs
@@ -190,7 +190,7 @@ mod test {
     }
 
     // Build a test stackframe
-    fn f(i: u32) -> StackFrame {
+    fn f(i: usize) -> StackFrame {
         StackFrame {
             name: format!("func{}", i),
             relative_path: format!("file{}.rb", i),
